+++ scripts/check_static_quality.sh
@@ -25,7 +25,7 @@ run_step "local resource intake docs" ./scripts/check_local_resource_intake.sh
 run_step "GitHub Project board docs" ./scripts/check_github_project_board_docs.sh
 run_step "ecosystem handoff docs" ./scripts/check_ecosystem_handoff.sh
 run_step "maintainer workflow docs" ./scripts/check_maintainer_workflow.sh
-run_step "issue templates" ./scripts/check_issue_templates.sh
+run_step "GitHub templates" ./scripts/check_issue_templates.sh
 run_step "README first steps" ./scripts/check_readme_first_steps.sh
 run_step "API reference coverage" ./scripts/check_api_reference.sh
 run_step "template manifest schema" ./scripts/check_template_manifest_schema.sh